feat: virtual keyboard for signs, lockers and more
This uses the SteamVR Overlays keyboard to enable editing of the games
input fields(lockers, beacons, vehicle names and signs).
It was tweaked to be as frictionless as possible for lockers and still
allow editing all the signs other features.
It also replaces the input dialogue of the beacon by the SteamVR one
completely since that one was ugly and not in world space anyways.

Additionaly it allows closing/deselecting input groups at all times
using UICancel. Previously you would have to point away from the input
group to close it, which I worked around for the CraftingMenu already.
Maybe that workaround is not needed anymore.

Closes #1